fix: changed int to long unsigned int

+++ oop-library/main.cpp
@@ -56,7 +56,7 @@ void LibraryUser::borrow(Item *item) {
 }
 
 void LibraryUser::return_item(Item *item) {
-  for (int i = 0; i < this->_borrowed_items.size(); i++) {
+  for (long unsigned int i = 0; i < this->_borrowed_items.size(); i++) {
     if (item == this->_borrowed_items[i]) {
       this->_borrowed_items.erase(this->_borrowed_items.begin() + i);
     }
@@ -64,7 +64,7 @@ void LibraryUser::return_item(Item *item) {
 }
 
 void LibraryUser::list_borrowed_items() {
-  for (int i = 0; i < this->_borrowed_items.size(); i++) {
+  for (long unsigned int i = 0; i < this->_borrowed_items.size(); i++) {
     std::cout << this->_borrowed_items[i] << std::endl;
   }
 }
@@ -85,11 +85,11 @@ class Library {
 };
 
 Library::~Library() {
-  for (int i = 0; i < this->_items.size(); i++) {
+  for (long unsigned int i = 0; i < this->_items.size(); i++) {
     this->_items.erase(this->_items.begin() + i);
   }
 
-  for (int i = 0; i < this->_users.size(); i++) {
+  for (long unsigned int i = 0; i < this->_users.size(); i++) {
     this->_users.erase(this->_users.begin() + i);
   }
 }
@@ -98,7 +98,7 @@ void Library::add_item(Item *item) {
 }
 
 void Library::remove_item(Item *item) {
-  for (int i = 0; i < this->_items.size(); i++) {
+  for (long unsigned int i = 0; i < this->_items.size(); i++) {
     if (item == this->_items[i]) {
       this->_items.erase(this->_items.begin() + i);
       break;
@@ -119,7 +119,7 @@ void Library::return_item(LibraryUser *user, Item *item) {
 }
 
 void Library::list_items() {
-  for (int i = 0; i < this->_items.size(); i++) {
+  for (long unsigned int i = 0; i < this->_items.size(); i++) {
     this->_items[i]->display_info();
   }
 }
